Fix Agent Suite parallel testing
While running ginkgo with the `-p` flag we were getting the error:

```
Ginkgo timed out waiting for all parallel nodes to report back!
```

The Agent Suite includes two tests that listen on ports, which we were
hardcoding to 1000. When run in serial it was fine because an AfterEach
would increase them, but in parallel they would overlap.

Increased the port variable by the GinkgoParallelNode number times 15,
which should give plenty of space for the two tests to not overlap.