fix: Increase scraper timeout
A `ScraperTimeout` of one second appears to be too short a timeout, as
we've seen instances (especially in more heavily loaded foundations)
where the scraper error logs include the following:
```
2023/09/18 17:45:18 failed to scrape: scrape errors:
[id: control, instance_id: , metric_url: https://10.0.4.9:53035/metrics]: Get "https://10.0.4.9:53035/metrics": context deadline exceeded (Client.Timeout exceeded while awaiting headers)
[id: control, instance_id: , metric_url: https://10.0.4.8:53035/metrics]: Get "https://10.0.4.8:53035/metrics": context deadline exceeded (Client.Timeout exceeded while awaiting headers)
```

Increasing it to one minute, the same as `ScrapeInterval` was somewhat
arbitrary, but we justified it by thinking: let the request run until
its time to scrape again.

While we were here, we also realized that this transport was an outdated
copy of `http.DefaultTransport`, so we updated it to use the latest
values.

Signed-off-by: Matthew Kocher <mkocher@vmware.com>
Signed-off-by: Carson Long <lcarson@vmware.com>